Stricter error handling for -Z
When we go to perform a ping for a particular AF check that a source
address was selected, if -Z was specified. If no source address was
selected then an error is printed and ping exits.

ie. when -Z is used we must find an address on that interface of the
same AF as the target, otherwise we error out.

VRVDR-40612