Correct -Z loopback/linklocal address checks
The logic used for the -Z option retrieves the addresses configured
on the specified interface and should only choose an address which
is not a loopback and, for IPv6, is not a link local address.

These checks were not checking the addresses which are assigned to
the interface and were therefore incorrectly passing, allowing a
source address to be selected when we did not intend it to be.

VRVDR-40612